Playlist thumbnail manager component
